+++ CHANGELOG.md
@@ -23,6 +23,10 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Updated file code to read from `data_static` as fallback in new location allowed in .gma (by @EntranceJew)
 - Scoreboard now sets preferred player volume and mute state in client's new `ttt2_voice` table (by @EntranceJew)
   - Keyed by steamid64, making it more reliable than UniqueID or the per-session mute and volume levels.
+- Added "olValue" to the onSelect/onChange function of comboboxes (by @NickCloudAT)
+
+### Fixed
+- Fixed Comboboxes throwing an invalid callback error in some instances (by @NickCloudAT)
 
 ## [v0.11.7b](https://github.com/TTT-2/TTT2/tree/v0.11.7b) (2022-08-27)